# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top Maracana API",
        "description": "Documentation de l'API Top Maracana",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/groups/{id}": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Détail d'un groupe",
                "operationId": "38a1f4990d0dc69ead7f7377ad016117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du groupe"
                    },
                    "404": {
                        "description": "Groupe introuvable"
                    }
                }
            }
        },
        "/api/groups/{id}/matches": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Matchs d'un groupe",
                "operationId": "fa95fd0ca9bd1100d8583c65c84872c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matchs du groupe"
                    },
                    "404": {
                        "description": "Groupe introuvable"
                    }
                }
            }
        },
...
        "/api/matches/{id}/goals": {
            "post": {
                "tags": [
                    "Matchs"
                ],
                "summary": "Ajouter un but",
                "operationId": "035e4e59666fff7c09af0cc86ca647ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "string",
                                        "example": "uuid-du-joueur"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "But ajouté avec succès"
                    },
                    "400": {
                        "description": "Erreur"
                    },
                    "404": {
                        "description": "Match ou joueur introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches/{id}/cards": {
            "post": {
                "tags": [
                    "Matchs"
                ],
                "summary": "Ajouter un carton",
                "operationId": "2cba1bb083acec622de5352ba755a4a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "card_type",
                                    "minute"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "string",
                                        "example": "uuid-du-joueur"
                                    },
                                    "card_type": {
                                        "type": "string",
                                        "enum": [
                                            "yellow",
                                            "red"
                                        ]
                                    },
                                    "minute": {
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carton ajouté avec succès"
                    },
                    "400": {
                        "description": "Erreur"
                    },
                    "404": {
                        "description": "Match ou joueur introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches/{id}/close": {
            "post": {
                "tags": [
                    "Matchs"
                ],
                "summary": "Clôturer un match",
                "operationId": "645fd8404ace221f2b4a7f5d0a90109c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match clôturé avec succès"
                    },
                    "400": {
                        "description": "Match déjà clôturé"
                    },
                    "404": {
                        "description": "Match introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches/{id}": {
            "get": {
                "tags": [
                    "Matchs"
                ],
                "summary": "Détail d'un match",
                "operationId": "496a55045627bb07d84440c12ba7f2b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du match"
                    },
                    "404": {
                        "description": "Match introuvable"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}